backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.5.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
prometheus-client==0.19.0
slowapi==0.1.9
python-dotenv==1.0.0
httpx==0.25.2
python-dateutil==2.8.2
email-validator==2.1.0
psutil==7.1.3

pypandoc==1.16.2